requirements.txt
fastapi
uvicorn
numpy
pillow
python-multipart
scikit-learn
joblib
scipy
requests
